feat: combine Grenzen & Achsen and Darstellung into one menu item
Merges the former section-range and section-settings editor sections
into a single SectionDisplay ("Grenzen & Darstellung") based on teacher
feedback. The combined hub card now appears as the first item above
"Objekte", and the tab strip shows the combined tab first.

closes #42
