--- projects/addin/src/app/components/plot-editor/sections/section-settings/section-settings.html
@@ -1,69 +1,0 @@
-@let form = store.editorForm;
-
-<div class="flex flex-col gap-2.5 p-2">
-  <section
-    class="border-default-medium flex flex-col gap-2 rounded-lg border bg-white p-2.5"
-  >
-    <h3 class="text-body-subtle text-2xs font-bold tracking-[.4px] uppercase">
-      Achsen
-    </h3>
-    <lg-toggle-row
-      [formField]="form.showAxis"
-      label="Achsen anzeigen"
-      sub="X- und Y-Achse als Linie"
-    />
-    @if (form.showAxis().value()) {
-      <lg-toggle-row
-        [formField]="form.showAxisLabels"
-        label="Achsenbeschriftung"
-        sub="Zahlenwerte an den Gitterlinien"
-      />
-      <lg-toggle-row
-        [formField]="form.showAxisArrows"
-        label="Pfeilspitzen"
-        sub="An den Enden der Achsen"
-      />
-    }
-  </section>
-
-  <section
-    class="border-default-medium flex flex-col gap-2 rounded-lg border bg-white p-2.5"
-  >
-    <h3 class="text-body-subtle text-2xs font-bold tracking-[.4px] uppercase">
-      Gitter
-    </h3>
-    <lg-toggle-row
-      [formField]="form.placeAxisLabelsInside"
-      label="Achsenbeschriftung an 0-Linie"
-      sub="Beschriftung in die Achsen schieben"
-    />
-    <div>
-      <span class="text-body-subtle text-2xs mb-1 block font-semibold"
-        >Schrittweite</span
-      >
-      <lg-pill-switch
-        [options]="gridStepOptions"
-        [formField]="form.gridStep"
-        aria-label="Schrittweite"
-        shape="square"
-      />
-    </div>
-  </section>
-
-  <section
-    class="border-default-medium flex flex-col gap-2 rounded-lg border bg-white p-2.5"
-  >
-    <h3 class="text-body-subtle text-2xs font-bold tracking-[.4px] uppercase">
-      Legende
-    </h3>
-    <div>
-      <span class="text-body-subtle text-2xs mb-1 block font-semibold"
-        >Legendenformat</span
-      >
-      <lg-dropdown
-        [formField]="form.legendLabelFormat"
-        [options]="legendLabelFormatOptions"
-      />
-    </div>
-  </section>
-</div>
