+++ projects/addin/src/app/components/settings/settings.html
@@ -14,7 +14,6 @@
     Deine Änderungen werden sofort gespeichert.
   </p>
 
-  <!-- Linien -->
   <section
     class="border-default-medium rounded-base flex flex-col gap-3 border bg-white p-3"
   >
@@ -83,7 +82,6 @@
     </div>
   </section>
 
-  <!-- Punkte -->
   <section
     class="border-default-medium rounded-base flex flex-col gap-3 border bg-white p-3"
   >
@@ -102,7 +100,6 @@
     </div>
   </section>
 
-  <!-- Legende -->
   <section
     class="border-default-medium rounded-base flex flex-col gap-3 border bg-white p-3"
   >
@@ -119,7 +116,6 @@
     </div>
   </section>
 
-  <!-- Vorschau -->
   <section
     class="border-default-medium rounded-base flex flex-col gap-3 border bg-white p-3"
   >
